+++ src/test/suite/configuration.test.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 import * as assert from 'assert';
-import { configuration, ConfigurationManager, ISetting, OutputAs } from '../../configuration';
+import { ConfigurationManager, ExecutionActionPreference, OutputAs } from '../../configuration';
 
 suite('ConfigurationManager tests', () => {
     test('Defaults', () => {
@@ -14,6 +14,8 @@ suite('ConfigurationManager tests', () => {
         assert.equal(config.get().documentationPath, undefined);
         assert.equal(config.get().documentationSnippet, 'Rule Doc');
         assert.equal(config.get().executionNotProcessedWarning, false);
+        assert.equal(config.get().executionRuleExcluded, ExecutionActionPreference.None);
+        assert.equal(config.get().executionRuleSuppressed, ExecutionActionPreference.None);
         assert.equal(config.get().experimentalEnabled, false);
         assert.equal(config.get().outputAs, OutputAs.Summary);
         assert.equal(config.get().notificationsShowChannelUpgrade, true);